fix(metastore): add index to support index name wildcard expansion (#5438)
The sql that is generated to expand an index name (index_id) into all of
the matching index names uses the LIKE clause to do partial text
matches. However, the only index on the index_id column utilizes the
default btree configuration which can only do exact value comparisons.
In environments with a large search volume over many thousands of
indexes, this can be a bottle neck as the query will always do an full
table scan. This drops and recreates the unique index on index_id
to include the `varchar_pattern_ops` which allows for wildcard
matching on a varchar column using the LIKE operator

Fixes: #5437