Bugfix sort order query string in elastic search api.
The sort order was parsed using the generated from_str_name method.
The spec of this method probably changed at one point, and we ended up
requiring sort order to be uppercase.

This PR changes the behavior to follow what we do when we parse a json
string, and adds unit tests.

Closes #5491