Prevents single split searches from different leaf_search from interleaving.
Before this PR, we just used a semaphore to acquire a permit and
start a new tokio task to run our single split search.

In pseudo code, a leaf_search would look like:
```
for split in splits {
    let permit = semaphore.acquire().await;
    tokio::spawn(async move {
       single_split_search(split);
       drop(permit)
    })
}
```

The problem with this is that it allows interleaving split search from
one search request with another one.

This interleaving strongly impacts search latency.
For instance, one can imagine two queries A and B with 3 splits each.
Executing as follows
| A | A | A | B | B | B |
offers a much short latency for A than
| A | B | B | A | B | A |

This PR also adds two metrics to monitor the number of queue single
split search.