Publish docker image only in "production" environment.
The point of this environment is that it now hosts the github secrets
that are required to push to github.

That environment is configured to require a CR from a short list
of 6 major contributors.

The goal here is to prevent a malicious contributor with write access
to access the docker api token secrets.