Removing the default field and body from the all field.
This PR is to be shipped in combination with
https://github.com/DataDog/pomsky-helm-charts/pull/17

In order to make pomsky lighter in storages,
we remove the body field from the concatenated all field,
and we remove the default field.

This PR adds logic so that we get the same result.
Basically queries targeting default (or all) previously
now target respectively error and message (respectively
all and message).