Use `pomsky` service account explicitly in staging
Explicitly configure the use of the `pomsky` service account
for staging deployments, instead of using the default generated
name. The default generated name changed from `pomsky` to `cloudprem`
due to recent chart updates, and to avoid breaking the mapping
back to S3 access, the pre-existing name needs to be used.

The intention is to follow this up with a change to the mapping in
the `cloud-inventory` repo to stnadardize on the `cloudprem` naming
convention.