Configure US1 staging to use orgstore for metastore
Configures the staging env to use an OrgStore Postgres database
for the metastore. This involves templating the credentials
via `consul-template` in a new init container, with the resulting
password being saved in a pgpass file that is mounted in a
memory volume.

This results in a relatively noisy values file, but is done with
minimal downstream impact on the shared charts. Peforming this
change entirely via the values file is chosen because this seems
to be a relatively custom setup, and so it is likely not worth
integrating it directly in the shared Helm charts.