TableView on group by term is broken in Log Explorer. (#206)
The reason is that we currently do not support exporting total count.
ES API does not have a direct equivalent, but it has a sum_other_doc_count
value at the aggregation level.

This PR populates the total metric when requested, and computes its value by summing
the returned buckets with sum_other_doc_count.

Closes #CLOUDPREM-185