Use the appropriate indexes storage URI for staging
The original staging setup had the indexes storage and metastore
storage reversed (i.e. the metastore storage was set to write
under /indexes), and this became apparent when switching to
OrgStore for the metastore.

This simply corrects the one remaining entry (for the index root
URI). It will result in current indexed data being orphaned in
staging, which seems totally acceptable, as it will avoid further
copy-paste of this erroneous setup.