Align staging resource requests with public documentation.
Update the resource requests for Pomsky in US1 staging to be
aligned with the suggested resource settings defined in the
public documentation (https://docs.datadoghq.com/cloudprem/configure/cluster_sizing).

Note that the lower-bound of suggested size is used for the
indexer (for now), and local storage and auto-scaling
are _not_ (yet) configured here.

Related to CLOUDPREM-181.